refactor block_explorer mod and add GemSwapExplorerInput
